requirements.txt
streamlit
pandas
Pillow
numpy
ultralytics
opencv-python
